Add SEO metadata and refactor client pages for App Router compatibility
