+++ app/dashboard/analytics/page.tsx
@@ -1,9 +1,16 @@
-import { Metadata } from 'next';
-import AnalyticsDashboard from '@/components/dashboard/analytics-dashboard';
+import { Metadata } from "next";
+import AnalyticsDashboard from "@/components/dashboard/analytics-dashboard";
 
 export const metadata: Metadata = {
   title: "Analytics | DraftDeckAI",
-  description: "View document usage statistics, engagement metrics, and improvement suggestions",
+  description:
+    "View document usage statistics, engagement metrics, and improvement suggestions",
+  openGraph: {
+    title: "Analytics | DraftDeckAI",
+    description:
+      "View document usage statistics, engagement metrics, and improvement suggestions",
+    type: "website",
+  },
 };
 
 /**
